Make sure user is logged in & in rides team to display admin settings
